chore: extra builder into its own file
