+++ libdd-data-pipeline/src/telemetry/mod.rs
@@ -2,10 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //! Telemetry provides a client to send results accumulated in 'Metrics'.
-pub mod error;
+pub(super) mod builder;
+pub(super) mod error;
 #[cfg(feature = "telemetry")]
-pub mod metrics;
-pub mod worker;
+mod metrics;
+pub(super) mod worker;
 
 use crate::telemetry::error::TelemetryError;
 #[cfg(feature = "telemetry")]
@@ -13,18 +14,13 @@ use crate::telemetry::metrics::Metrics;
 #[cfg(feature = "telemetry")]
 use libdd_common::tag::Tag;
 #[cfg(feature = "telemetry")]
-use libdd_telemetry::worker::{
-    LifecycleAction, TelemetryActions, TelemetryWorkerBuilder, TelemetryWorkerFlavor,
-    TelemetryWorkerHandle,
-};
+use libdd_telemetry::worker::{LifecycleAction, TelemetryActions, TelemetryWorkerHandle};
 #[cfg(feature = "telemetry")]
 use libdd_trace_utils::send_with_retry::SendWithRetryError;
 use libdd_trace_utils::send_with_retry::SendWithRetryResult;
 use libdd_trace_utils::trace_utils::SendDataResult;
-#[cfg(feature = "telemetry")]
-use std::collections::HashMap;
-use std::time::Duration;
-use tokio::runtime::Handle;
+
+pub use builder::TelemetryClientBuilder;
 
 /// Configuration for telemetry reporting.
 #[derive(Debug, Default, Clone)]
@@ -32,137 +28,6 @@ pub struct TelemetryConfig {
     pub heartbeat: u64,
     pub runtime_id: Option<String>,
     pub debug_enabled: bool,
-}
-
-/// Structure to build a Telemetry client.
-///
-/// Holds partial data until the `build` method is called which results in a new
-/// `TelemetryClient`.
-#[derive(Default)]
-pub struct TelemetryClientBuilder {
-    service_name: Option<String>,
-    service_version: Option<String>,
-    env: Option<String>,
-    language: Option<String>,
-    language_version: Option<String>,
-    tracer_version: Option<String>,
-    url: Option<String>,
-    heartbeat: Option<Duration>,
-    debug_enabled: bool,
-    runtime_id: Option<String>,
-}
-
-impl TelemetryClientBuilder {
-    /// Sets the service name for the telemetry client
-    pub fn set_service_name(mut self, name: &str) -> Self {
-        self.service_name = Some(name.to_string());
-        self
-    }
-
-    /// Sets the service version for the telemetry client
-    pub fn set_service_version(mut self, version: &str) -> Self {
-        self.service_version = Some(version.to_string());
-        self
-    }
-
-    /// Sets the env name for the telemetry client
-    pub fn set_env(mut self, name: &str) -> Self {
-        self.env = Some(name.to_string());
-        self
-    }
-
-    /// Sets the language name for the telemetry client
-    pub fn set_language(mut self, lang: &str) -> Self {
-        self.language = Some(lang.to_string());
-        self
-    }
-
-    /// Sets the language version for the telemetry client
-    pub fn set_language_version(mut self, version: &str) -> Self {
-        self.language_version = Some(version.to_string());
-        self
-    }
-
-    /// Sets the tracer version for the telemetry client
-    pub fn set_tracer_version(mut self, version: &str) -> Self {
-        self.tracer_version = Some(version.to_string());
-        self
-    }
-
-    /// Sets the url where the metrics will be sent.
-    pub fn set_url(mut self, url: &str) -> Self {
-        self.url = Some(url.to_string());
-        self
-    }
-
-    /// Sets the heartbeat notification interval in millis.
-    pub fn set_heartbeat(mut self, interval: u64) -> Self {
-        if interval > 0 {
-            self.heartbeat = Some(Duration::from_millis(interval));
-        }
-        self
-    }
-
-    /// Sets runtime id for the telemetry client.
-    pub fn set_runtime_id(mut self, id: &str) -> Self {
-        self.runtime_id = Some(id.to_string());
-        self
-    }
-
-    /// Sets the debug enabled flag for the telemetry client.
-    pub fn set_debug_enabled(mut self, debug: bool) -> Self {
-        self.debug_enabled = debug;
-        self
-    }
-}
-
-#[cfg(feature = "telemetry")]
-impl TelemetryClientBuilder {
-    /// Builds the telemetry client.
-    pub fn build(self, runtime: Handle) -> (TelemetryClient, worker::TelemetryWorker) {
-        #[allow(clippy::unwrap_used)]
-        let mut builder = TelemetryWorkerBuilder::new_fetch_host(
-            self.service_name.unwrap(),
-            self.language.unwrap(),
-            self.language_version.unwrap(),
-            self.tracer_version.unwrap(),
-        );
-        if let Some(url) = self.url {
-            let _ = builder
-                .config
-                .set_endpoint(libdd_common::Endpoint::from_slice(&url));
-        }
-        if let Some(heartbeat) = self.heartbeat {
-            builder.config.telemetry_heartbeat_interval = heartbeat;
-        }
-        builder.config.debug_enabled = self.debug_enabled;
-        // Send only metrics and logs and drop lifecycle events
-        builder.flavor = TelemetryWorkerFlavor::MetricsLogs;
-        builder.application.env = self.env;
-        builder.application.service_version = self.service_version;
-
-        if let Some(id) = self.runtime_id {
-            builder.runtime_id = Some(id);
-        }
-
-        let (worker_handle, worker) = builder.build_worker(runtime);
-
-        (
-            TelemetryClient {
-                metrics: Metrics::new(&worker_handle),
-                worker: worker_handle,
-            },
-            worker,
-        )
-    }
-}
-
-#[cfg(not(feature = "telemetry"))]
-impl TelemetryClientBuilder {
-    /// Builds a no-op telemetry client.
-    pub fn build(self, _runtime: Handle) -> (TelemetryClient, worker::TelemetryWorker) {
-        (TelemetryClient {}, worker::TelemetryWorker {})
-    }
 }
 
 #[cfg(feature = "telemetry")]
@@ -288,7 +153,7 @@ pub struct SendPayloadTelemetry {
     chunks_dropped_p0: u64,
     chunks_dropped_serialization_error: u64,
     chunks_dropped_send_failure: u64,
-    responses_count_per_code: HashMap<u16, u64>,
+    responses_count_per_code: std::collections::HashMap<u16, u64>,
 }
 
 #[cfg(not(feature = "telemetry"))]
@@ -397,7 +262,9 @@ mod tests {
     use httpmock::MockServer;
     use libdd_common::{http_common, worker::Worker};
     use regex::Regex;
-    use tokio::time::sleep;
+    use std::collections::HashMap;
+    use std::time::Duration;
+    use tokio::{runtime::Handle, time::sleep};
 
     use super::*;
 
@@ -415,30 +282,6 @@ mod tests {
             .build(Handle::current());
         tokio::spawn(async move { worker.run().await });
         client
-    }
-
-    #[test]
-    fn builder_test() {
-        let builder = TelemetryClientBuilder::default()
-            .set_service_name("test_service")
-            .set_service_version("test_version")
-            .set_env("test_env")
-            .set_language("test_language")
-            .set_language_version("test_language_version")
-            .set_tracer_version("test_tracer_version")
-            .set_url("http://localhost")
-            .set_debug_enabled(true)
-            .set_heartbeat(30);
-
-        assert_eq!(&builder.service_name.unwrap(), "test_service");
-        assert_eq!(&builder.service_version.unwrap(), "test_version");
-        assert_eq!(&builder.env.unwrap(), "test_env");
-        assert_eq!(&builder.language.unwrap(), "test_language");
-        assert_eq!(&builder.language_version.unwrap(), "test_language_version");
-        assert_eq!(&builder.tracer_version.unwrap(), "test_tracer_version");
-        assert!(builder.debug_enabled);
-        assert_eq!(builder.url.as_deref(), Some("http://localhost"));
-        assert_eq!(builder.heartbeat.unwrap(), Duration::from_millis(30));
     }
 
     #[cfg_attr(miri, ignore)]
